refactor: move the release subsystem from runner.py into release.py (Issue #286)
Pure move of the release delivery state machine (declaration parsing,
scope verification, gates, version preparation, tag, GitHub Release
publish, docs sync, Milestone close, comment bodies, process_release)
into the new src/orbi/release.py. release imports the runner
primitives; the dispatch in process_issue imports release lazily, so
no circular import and the runner-side inbound surface stays the
single process_issue dispatch. No logic changes.

The Issue's stale function list is adapted to the current tree: four
of the named functions no longer exist (removed by later work) and the
release-only neighbors the named functions call (derive scope, prepare
version, docs sync, docs page helpers) move with the block; _is_ancestor
and the epic/milestone reconciliation family stay in runner (tick and
merge-gate callers). Tests change import/patch paths only.
