+++ tests/conftest.py
@@ -25,8 +25,16 @@ def _default_cli_install_preflight(monkeypatch):
     always wins over this default). The implementation lives in
     `orbi.runner` itself (see the NOTE there), so the stub
     patches ITS module global — the call `main()` makes."""
+    import orbi.release as release
+
     monkeypatch.setattr(
         runner, "refresh_cli_install", lambda *a, **k: "unchanged",
+    )
+    # Issue #286: `process_release` now lives in `orbi.release` and reads
+    # the refresh through the release module global — stub that binding
+    # with the same default no-op.
+    monkeypatch.setattr(
+        release, "refresh_cli_install", lambda *a, **k: "unchanged",
     )
